requirements.txt
'''
streamlit
groq
sentence-transformers
faiss-cpu
langchain
langchain-community
PyPDF2
requests
numpy